scripts/requirements.txt
numpy>=1.21.0
sounddevice>=0.4.5
scipy>=1.7.0
pydub>=0.25.1
openai-whisper
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.27.0 